# openapi/spec.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Subscriptions Service
  description: A service that deals with access to the KL platform
paths:
...
  /subscriptions:
    post:
      description: Create a Subscription
      x-auth-api-key: create-subscription
      requestBody:
        $ref: "#/components/requestBodies/CreateSubscriptionRequest"
      responses:
        "201":
          description: "Successfully created a Subscription"
          headers:
            Location:
              schema:
                type: string
                format: uri
                description: "Location of the created resource"
        "409":
          description: "Could not create Subscription because one already exists for the Account"
        "401":
          description: "The API key provided is not recognised"
        "403":
          description: "The API key provided is not allowed to call this endpoint"
    get:
      description: Search for a Subscription
      x-auth-api-key: get-subscription
      x-auth-jwt: true
      parameters:
        - name: account_id
          schema:
            type: string
          in: query
          required: true
      responses:
        "200":
          description: "Results from search for Subscription"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subscription"
        "401":
          description: "The API key provided is not recognised"
        "403":
          description: "The API key provided is not allowed to call this endpoint"
...
components:
  schemas:
...
    Subscription:
      required:
        - id
        - account_id
        - state
      properties:
        id:
          type: string
          format: uuid
        account_id:
          type: string
          format: uuid
        state:
          type: string
...
  requestBodies:
    CreateSubscriptionRequest:
      description: Request to create a new Subscription
      required: true
      content:
        application/json:
          schema:
            type: object
            required:
              - account_id
              - state
            properties:
              account_id:
                type: string
                format: uuid
...